Add article author and date, improve article layout

+++ Source/PlayerState.swift
@@ -11,7 +11,7 @@ public struct PlayerState {
     public var visible: Bool = true
     public var maximised: Bool = false
     public var playing: Bool = false
-    public var position: Float = 0
+    public var position: Double = 0
 
     public var title: String? = nil
     public var body: String? = nil
